WIP: More logging restructuring
- remove custom*Error lambda gymnastics and rely on logging
- code currently not functional

 On branch logging_enh
 Changes to be committed:
	modified:   acme/dask_helpers.py
	modified:   acme/spy_interface.py
